Added a spotting method call.
Contains a method, with JavaDoc, that can be called to return a list of
any players that are in the casters line of sight. Returns a bukkit
Player, but could be modified to return SpoutPlayers, or a custom
extension.

The function does nothing upon discovering the spot, simply confirms
its existence.